refactor: rename validation js file
